+++ code/core/src/main/java/io/zabrek/soulbound/SoulBound.java
@@ -1,18 +1,18 @@
 package io.zabrek.soulbound;
 
-import io.zabrek.soulbound.api.kernel.CoreComponent;
 import io.zabrek.soulbound.api.kernel.CoreComponentLoader;
+import io.zabrek.soulbound.api.logger.SoulBoundLogger;
+import io.zabrek.soulbound.api.logger.SoulBoundLoggerFactory;
 import io.zabrek.soulbound.kernel.SoulBoundComponents;
 import io.zabrek.soulbound.kernel.TopologicalCoreComponentLoader;
+import io.zabrek.soulbound.lib.logger.CachingSoulBoundLoggerFactory;
+import io.zabrek.soulbound.logger.DefaultSoulBoundLoggerFactory;
 import org.bukkit.Server;
 import org.bukkit.plugin.PluginDescriptionFile;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.ServicesManager;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitScheduler;
-
-import java.util.Set;
-import java.util.logging.Logger;
 
 /**
  * Represents SoulBound plugin.
@@ -25,9 +25,9 @@ public class SoulBound extends JavaPlugin {
     private static SoulBound instance;
 
     /**
-     * The logger instance.
+     * The custom logger for the plugin.
      */
-    protected Logger log;
+    protected SoulBoundLogger log;
 
     /**
      * The loader responsible for discovering and registering core components.
@@ -41,19 +41,23 @@ public class SoulBound extends JavaPlugin {
         super();
     }
 
+    /**
+     * Get the plugin's instance.
+     *
+     * @return The plugin's instance.
+     */
+    public static SoulBound getInstance() {
+        return instance;
+    }
+
     @Override
     public void onEnable() {
         instance = this;
 
-        final Set<CoreComponent> defaults = SoulBoundComponents.createDefaults(this);
-        for (final CoreComponent component : defaults) {
-            loader.register(component);
-        }
-
         try {
             loader.load();
         } catch (final Exception e) {
-            log.severe("Failed to load SoulBound components: " + e.getMessage());
+            log.error("Failed to load SoulBound components: " + e.getMessage());
             getServer().getPluginManager().disablePlugin(this);
             return;
         }
@@ -63,10 +67,13 @@ public class SoulBound extends JavaPlugin {
 
     @Override
     public void onLoad() {
-        this.log = getLogger();
+        final SoulBoundLoggerFactory loggerFactory = new CachingSoulBoundLoggerFactory(new DefaultSoulBoundLoggerFactory());
+        this.log = loggerFactory.create(this);
 
         this.loader = new TopologicalCoreComponentLoader();
+        this.loader.init(SoulBoundLoggerFactory.class, loggerFactory);
         initPluginDependencies(loader);
+        SoulBoundComponents.createDefaults(this).forEach(loader::register);
     }
 
     @Override
@@ -81,15 +88,6 @@ public class SoulBound extends JavaPlugin {
         loader.init(BukkitScheduler.class, getServer().getScheduler());
         loader.init(PluginDescriptionFile.class, getDescription());
         loader.init(ServicesManager.class, getServer().getServicesManager());
-    }
-
-    /**
-     * Get the plugin's instance.
-     *
-     * @return The plugin's instance.
-     */
-    public static SoulBound getInstance() {
-        return instance;
     }
 
     /**