+++ code/lib/src/main/java/io/zabrek/soulbound/lib/logger/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains default implementations related to logging.
+ */
+package io.zabrek.soulbound.lib.logger;